geotiff: extract GPU helpers to _backends/_gpu_helpers.py (#1884)
Step 6 of #1813's multi-PR refactor of __init__.py. Pure code motion;
no public API change. First PR to create the _backends/ subpackage.

Created xrspatial/geotiff/_backends/__init__.py (empty placeholder)
and moved into _backends/_gpu_helpers.py:

- _is_gpu_data: cupy-backed array / DataArray detection.
- _apply_nodata_mask_gpu: cupy mirror of the CPU sentinel-to-NaN mask
  (handles float and integer arrays, with the same out-of-range /
  fractional / non-finite skips as _resolve_masked_fill in
  _reader.py).
- _gpu_decode_single_band_tiles: two-stage GPU decode (GDS then
  CPU-extracted bytes) for the PlanarConfiguration=2 path.
- _apply_orientation_gpu: cupy mirror of the CPU orientation-flip
  helper for the eight TIFF Orientation tag values.
- _apply_orientation_geo_info: GeoTransform update after an
  orientation flip, with the #1765 refusal for georeferenced
  orientations 5-8.
- _gpu_apply_window_band: post-decode window + band slice on device.

Every helper lazy-imports cupy at call time so the module imports
cleanly on numpy-only environments. __init__.py re-imports the
helpers from _backends._gpu_helpers so existing callers
(read_geotiff_gpu, to_geotiff's GPU branch, write_geotiff_gpu) keep
their identical signatures.

Leading-underscore module name (_gpu_helpers.py) keeps _backends/gpu.py
free for the read_geotiff_gpu entry-point body in step 7.

Net __init__.py change: 4271 -> 3969 lines (-302). _gpu_helpers.py is
335 lines.

Verification: pixel-parity matrix from #1889 (192 cells, all GPU
cells included), runtime identity (9), GPU byteswap (#1508), overview
nodata inheritance (#1739) -- 230/230 pass.

Refs #1813.
